+++ UnitySymbol/UnitySymbol.cs
@@ -7,8 +7,10 @@ namespace Hananoki {
 		static List<string> s_lst;
 
 		public static bool Has( string symbol ) {
-			if( s_lst == null) s_lst = EditorUserBuildSettings.activeScriptCompilationDefines.Where( x => x.StartsWith( "UNITY_" ) ).ToList();
+			if( s_lst == null ) s_lst = EditorUserBuildSettings.activeScriptCompilationDefines.Where( x => x.StartsWith( "UNITY_" ) ).ToList();
 			return s_lst.Contains( symbol );
 		}
+		public static bool UNITY_2019_1_OR_NEWER => Has( "UNITY_2019_1_OR_NEWER" );
+
 	}
 }